refactor: split config.rs into per-registry config modules (#484)
Break the 4639-line monolithic config.rs into 25 focused modules:
- config/mod.rs — orchestration, shared defaults, parse_env_warn, tests
- config/{server,storage,auth,rate_limit,gc,retention,curation,
  circuit_breaker,audit_cfg,registries}.rs — top-level subsystems
- config/registry/{npm,pypi,cargo,go,docker,maven,gems,raw,
  terraform,ansible,nuget,pub_dart,conan}.rs — per-format configs

Also fixes two pre-existing antipattern budget overflows:
- mirror/docker.rs: typed PushError instead of Result<_, String> (RI-17)
- registry/pub_dart.rs: tracing::debug on error instead of erasure (RI-19)

Updates coherence-check.sh to scan config/ directory instead of config.rs.

All 1204 tests pass. No user-facing config changes.
