+++ nora-registry/src/config/mod.rs
@@ -1,21 +1,99 @@
 // Copyright (c) 2026 The NORA Authors
 // SPDX-License-Identifier: MIT
 
+//! NORA registry configuration module.
+//!
+//! Configuration is loaded with three-tier priority: ENV > config file > defaults.
+//! Each subsystem (server, storage, auth, registries, etc.) is defined in its own
+//! submodule for maintainability.
+
+mod audit_cfg;
+mod auth;
+mod circuit_breaker;
+mod curation;
+mod gc;
+mod rate_limit;
+mod registries;
+pub mod registry;
+mod retention;
+mod server;
+mod storage;
+
+// Infrastructure configs
+pub use self::audit_cfg::AuditConfig;
+// Re-exports maintain API surface: `crate::config::OidcRoleRule` etc. used by test code in auth/, circuit_breaker/
+#[allow(unused_imports)]
+pub use self::auth::{AuthConfig, OidcConfig, OidcProvider, OidcRoleRule, TrustedProxies};
+#[allow(unused_imports)]
+pub use self::circuit_breaker::{CircuitBreakerConfig, CircuitBreakerOverride};
+pub use self::curation::{
+    CurationConfig, CurationMode, CurationOnFailure, RegistryCurationOverride,
+};
+pub use self::gc::GcConfig;
+pub use self::rate_limit::RateLimitConfig;
+pub use self::registries::{EnableSpec, RegistriesSection};
+pub use self::retention::{RetentionConfig, RetentionRule};
+pub use self::server::{ServerConfig, TlsConfig};
+pub use self::storage::{StorageConfig, StorageMode};
+
+// Registry configs (re-exported from registry/ submodule tree)
+pub use self::registry::*;
+
+// Secrets config lives in crate::secrets — just re-export
+pub use crate::secrets::SecretsConfig;
+
 use base64::{engine::general_purpose::STANDARD, Engine};
 use serde::{Deserialize, Serialize};
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 use std::env;
 use std::fs;
 
-use crate::digest_quarantine::QuarantineMode;
 use crate::registry_type::RegistryType;
 
-use crate::secrets::ProtectedString;
-pub use crate::secrets::SecretsConfig;
+// --- Shared defaults (used by submodules via `super::default_*`) ---
+
+/// Returns `true` — used as serde default for boolean fields that should default to enabled.
+pub(crate) fn default_true() -> bool {
+    true
+}
+
+/// Default proxy timeout in seconds (30s).
+pub(crate) fn default_timeout() -> u64 {
+    30
+}
+
+/// Default metadata TTL in seconds (300s = 5 minutes).
+pub(crate) fn default_metadata_ttl() -> i64 {
+    300
+}
 
 /// Encode "user:pass" into a Basic Auth header value, e.g. "Basic dXNlcjpwYXNz".
 pub fn basic_auth_header(credentials: &str) -> String {
     format!("Basic {}", STANDARD.encode(credentials))
+}
+
+/// Parse an env var value into `target`, warning on failure (#537).
+///
+/// If `val` parses successfully, `*target` is updated. If parsing fails,
+/// a `warn!` is emitted with the variable name and invalid value, and
+/// `*target` is left unchanged (keeps the TOML/default value).
+pub(crate) fn parse_env_warn<T: std::str::FromStr + std::fmt::Display>(
+    name: &str,
+    val: &str,
+    target: &mut T,
+) {
+    match val.parse::<T>() {
+        Ok(parsed) => {
+            *target = parsed;
+        }
+        Err(_) => {
+            tracing::warn!(
+                var = name,
+                value = val,
+                "env override ignored: failed to parse value"
+            );
+        }
+    }
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -69,1526 +147,6 @@ pub struct Config {
     /// Declarative registry selection: `[registries] enable = ["docker", "npm"]`
     #[serde(default)]
     pub registries: Option<RegistriesSection>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ServerConfig {
-    pub host: String,
-    pub port: u16,
-    /// Public URL for generating pull commands (e.g., "registry.example.com")
-    #[serde(default)]
-    pub public_url: Option<String>,
-    /// Maximum request body size in MB (default: 2048 = 2GB)
-    #[serde(default = "default_body_limit_mb")]
-    pub body_limit_mb: usize,
-}
-
-fn default_body_limit_mb() -> usize {
-    2048 // 2GB - enough for any Docker image
-}
-
-/// TLS configuration for outbound connections to upstream registries.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct TlsConfig {
-    /// Path to PEM-encoded CA certificate bundle (appended to system CAs)
-    #[serde(default)]
-    pub ca_cert: Option<String>,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "lowercase")]
-pub enum StorageMode {
-    #[default]
-    Local,
-    S3,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct StorageConfig {
-    #[serde(default)]
-    pub mode: StorageMode,
-    #[serde(default = "default_storage_path")]
-    pub path: String,
-    #[serde(default = "default_s3_url")]
-    pub s3_url: String,
-    #[serde(default = "default_bucket")]
-    pub bucket: String,
-    /// S3 access key (optional, uses anonymous access if not set)
-    #[serde(default, skip_serializing)]
-    pub s3_access_key: Option<ProtectedString>,
-    /// S3 secret key (optional, uses anonymous access if not set)
-    #[serde(default, skip_serializing)]
-    pub s3_secret_key: Option<ProtectedString>,
-    /// S3 region (default: us-east-1)
-    #[serde(default = "default_s3_region")]
-    pub s3_region: String,
-}
-
-fn default_s3_region() -> String {
-    "us-east-1".to_string()
-}
-
-fn default_storage_path() -> String {
-    "data/storage".to_string()
-}
-
-fn default_s3_url() -> String {
-    "http://127.0.0.1:9000".to_string()
-}
-
-fn default_bucket() -> String {
-    "registry".to_string()
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct MavenConfig {
-    #[serde(default = "default_true")]
-    pub enabled: bool,
-    #[serde(default)]
-    pub proxies: Vec<MavenProxyEntry>,
-    #[serde(default = "default_timeout")]
-    pub proxy_timeout: u64,
-    /// Verify client-uploaded checksums against server-computed values
-    #[serde(default = "default_true")]
-    pub checksum_verify: bool,
-    /// Prevent overwriting released (non-SNAPSHOT) artifacts
-    #[serde(default = "default_true")]
-    pub immutable_releases: bool,
-}
-
-fn default_true() -> bool {
-    true
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct NpmConfig {
-    #[serde(default = "default_true")]
-    pub enabled: bool,
-    #[serde(default)]
-    pub proxy: Option<String>,
-    #[serde(default, skip_serializing)]
-    pub proxy_auth: Option<ProtectedString>,
-    #[serde(default = "default_timeout")]
-    pub proxy_timeout: u64,
-    /// Metadata cache TTL in seconds (default: 300 = 5 min).
-    /// -1 = cache forever, 0 = always refetch, >0 = seconds.
-    #[serde(default = "default_metadata_ttl")]
-    pub metadata_ttl: i64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PypiConfig {
-    #[serde(default = "default_true")]
-    pub enabled: bool,
-    #[serde(default)]
-    pub proxy: Option<String>,
-    #[serde(default, skip_serializing)]
-    pub proxy_auth: Option<ProtectedString>,
-    #[serde(default = "default_timeout")]
-    pub proxy_timeout: u64,
-}
-
-/// Cargo registry configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CargoConfig {
-    #[serde(default = "default_true")]
-    pub enabled: bool,
-    /// Upstream Cargo registry (crates.io API)
-    #[serde(default = "default_cargo_proxy")]
-    pub proxy: Option<String>,
-    #[serde(default, skip_serializing)]
-    pub proxy_auth: Option<ProtectedString>,
-    #[serde(default = "default_timeout")]
-    pub proxy_timeout: u64,
-}
-
-fn default_cargo_proxy() -> Option<String> {
-    Some("https://crates.io".to_string())
-}
-
-impl Default for CargoConfig {
-    fn default() -> Self {
-        Self {
-            enabled: true,
-            proxy: default_cargo_proxy(),
-            proxy_auth: None,
-            proxy_timeout: 30,
-        }
-    }
-}
-
-/// Go module proxy configuration (GOPROXY protocol)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GoConfig {
-    #[serde(default = "default_true")]
-    pub enabled: bool,
-    /// Upstream Go module proxy URL (default: https://proxy.golang.org)
-    #[serde(default = "default_go_proxy")]
-    pub proxy: Option<String>,
-    #[serde(default, skip_serializing)]
-    pub proxy_auth: Option<ProtectedString>,
-    #[serde(default = "default_timeout")]
-    pub proxy_timeout: u64,
-    /// Separate timeout for .zip downloads (default: 120s, zips can be large)
-    #[serde(default = "default_go_zip_timeout")]
-    pub proxy_timeout_zip: u64,
-    /// Maximum module zip size in bytes (default: 100MB)
-    #[serde(default = "default_go_max_zip_size")]
-    pub max_zip_size: u64,
-}
-
-fn default_go_proxy() -> Option<String> {
-    Some("https://proxy.golang.org".to_string())
-}
-
-fn default_go_zip_timeout() -> u64 {
-    120
-}
-
-fn default_go_max_zip_size() -> u64 {
-    104_857_600 // 100MB
-}
-
-impl Default for GoConfig {
-    fn default() -> Self {
-        Self {
-            enabled: true,
-            proxy: default_go_proxy(),
-            proxy_auth: None,
-            proxy_timeout: 30,
-            proxy_timeout_zip: 120,
-            max_zip_size: 104_857_600,
-        }
-    }
-}
-
-/// Docker registry configuration with upstream proxy support
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DockerConfig {
-    #[serde(default = "default_true")]
-    pub enabled: bool,
-    #[serde(default = "default_docker_timeout")]
-    pub proxy_timeout: u64,
-    /// Per-chunk read timeout in seconds for streaming blob downloads (default: 60).
-    /// Applies to each chunk individually, not the total transfer.
-    #[serde(default = "default_docker_read_timeout")]
-    pub read_timeout: u64,
-    /// Metadata cache TTL in seconds (default: -1 = cache forever).
-    /// -1 = cache forever, 0 = always refetch, >0 = seconds.
-    #[serde(default = "default_docker_metadata_ttl")]
-    pub metadata_ttl: i64,
-    /// Serve stale cached manifests when upstream is unreachable (default: true).
-    #[serde(default = "default_true")]
-    pub serve_stale: bool,
-    #[serde(default)]
-    pub upstreams: Vec<DockerUpstream>,
-}
-
-/// Docker upstream registry configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DockerUpstream {
-    pub url: String,
-    #[serde(default, skip_serializing)]
-    pub auth: Option<ProtectedString>,
-    /// Storage namespace prefix (e.g. "docker.io"). Derived from URL host if omitted.
-    #[serde(default)]
-    pub namespace: Option<String>,
-    /// Path-based routing prefix (e.g. "docker-hub"). When set, requests to
-    /// `/v2/docker-hub/library/nginx/manifests/latest` route to this upstream
-    /// with the prefix stripped (name becomes `library/nginx`).
-    #[serde(default)]
-    pub prefix: Option<String>,
-}
-
-impl DockerUpstream {
-    /// Resolved namespace: explicit config or derived from URL host.
-    pub fn resolved_namespace(&self) -> String {
-        if let Some(ref ns) = self.namespace {
-            return ns.clone();
-        }
-        // Derive from URL host: "https://registry-1.docker.io" → "docker.io"
-        extract_docker_namespace(&self.url)
-    }
-}
-
-/// Derive a storage namespace from a Docker registry URL.
-///
-/// Strips scheme and common prefixes:
-/// - `https://registry-1.docker.io` → `docker.io`
-/// - `https://ghcr.io` → `ghcr.io`
-/// - `https://registry.example.com` → `example.com`
-pub fn extract_docker_namespace(url: &str) -> String {
-    let host = url
-        .strip_prefix("https://")
-        .or_else(|| url.strip_prefix("http://"))
-        .unwrap_or(url)
-        .split('/')
-        .next()
-        .unwrap_or("default")
-        .split(':')
-        .next()
-        .unwrap_or("default");
-    let host = host.strip_prefix("registry-1.").unwrap_or(host);
-    let host = host.strip_prefix("registry.").unwrap_or(host);
-    host.to_string()
-}
-
-/// Maven upstream proxy configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum MavenProxyEntry {
-    Simple(String),
-    Full(MavenProxy),
-}
-
-/// Maven upstream proxy with optional auth
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct MavenProxy {
-    pub url: String,
-    #[serde(default, skip_serializing)]
-    pub auth: Option<ProtectedString>,
-}
-
-impl MavenProxyEntry {
-    pub fn url(&self) -> &str {
-        match self {
-            MavenProxyEntry::Simple(s) => s,
-            MavenProxyEntry::Full(p) => &p.url,
-        }
-    }
-    pub fn auth(&self) -> Option<&str> {
-        use crate::secrets::expose_opt;
-        match self {
-            MavenProxyEntry::Simple(_) => None,
-            MavenProxyEntry::Full(p) => expose_opt(&p.auth),
-        }
-    }
-}
-
-/// Raw repository configuration for simple file storage
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RawConfig {
-    #[serde(default = "default_raw_enabled")]
-    pub enabled: bool,
-    #[serde(default = "default_max_file_size")]
-    pub max_file_size: u64, // in bytes
-    #[serde(default = "default_raw_cache_control")]
-    pub cache_control: String,
-}
-
-/// RubyGems proxy configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GemsConfig {
-    #[serde(default)]
-    pub enabled: bool,
-    /// Upstream RubyGems registry (default: https://rubygems.org)
-    #[serde(default = "default_gems_proxy")]
-    pub proxy: Option<String>,
-    #[serde(default, skip_serializing)]
-    pub proxy_auth: Option<ProtectedString>,
-    #[serde(default = "default_timeout")]
-    pub proxy_timeout: u64,
-    /// Metadata cache TTL in seconds (default: 300 = 5 min).
-    /// -1 = cache forever, 0 = always refetch, >0 = seconds.
-    #[serde(default = "default_metadata_ttl")]
-    pub metadata_ttl: i64,
-}
-
-fn default_gems_proxy() -> Option<String> {
-    Some("https://rubygems.org".to_string())
-}
-
-impl Default for GemsConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            proxy: default_gems_proxy(),
-            proxy_auth: None,
-            proxy_timeout: 30,
-            metadata_ttl: 300,
-        }
-    }
-}
-
-/// Terraform provider/module proxy configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TerraformConfig {
-    #[serde(default)]
-    pub enabled: bool,
-    /// Upstream Terraform registry (default: https://registry.terraform.io)
-    #[serde(default = "default_terraform_proxy")]
-    pub proxy: Option<String>,
-    #[serde(default, skip_serializing)]
-    pub proxy_auth: Option<ProtectedString>,
-    #[serde(default = "default_timeout")]
-    pub proxy_timeout: u64,
-    /// Separate timeout for binary downloads (default: 120s)
-    #[serde(default = "default_go_zip_timeout")]
-    pub proxy_timeout_dl: u64,
-    /// Metadata cache TTL in seconds (default: 300 = 5 min).
-    /// -1 = cache forever, 0 = always refetch, >0 = seconds.
-    #[serde(default = "default_metadata_ttl")]
-    pub metadata_ttl: i64,
-    /// Serve stale cached metadata when upstream is unreachable (default: true).
-    #[serde(default = "default_true")]
-    pub serve_stale: bool,
-}
-
-fn default_terraform_proxy() -> Option<String> {
-    Some("https://registry.terraform.io".to_string())
-}
-
-impl Default for TerraformConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            proxy: default_terraform_proxy(),
-            proxy_auth: None,
-            proxy_timeout: 30,
-            proxy_timeout_dl: 120,
-            metadata_ttl: 300,
-            serve_stale: true,
-        }
-    }
-}
-
-/// Ansible Galaxy collection proxy configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AnsibleConfig {
-    #[serde(default)]
-    pub enabled: bool,
-    /// Upstream Galaxy server (default: https://galaxy.ansible.com)
-    #[serde(default = "default_ansible_proxy")]
-    pub proxy: Option<String>,
-    #[serde(default, skip_serializing)]
-    pub proxy_auth: Option<ProtectedString>,
-    #[serde(default = "default_timeout")]
-    pub proxy_timeout: u64,
-    /// Metadata cache TTL in seconds (default: 3600 = 1 hour).
-    /// -1 = cache forever, 0 = always refetch, >0 = seconds.
-    #[serde(default = "default_ansible_metadata_ttl")]
-    pub metadata_ttl: i64,
-    /// Serve stale cached metadata when upstream is unreachable (default: true).
-    #[serde(default = "default_true")]
-    pub serve_stale: bool,
-}
-
-fn default_ansible_proxy() -> Option<String> {
-    Some("https://galaxy.ansible.com".to_string())
-}
-
-fn default_ansible_metadata_ttl() -> i64 {
-    3600
-}
-
-impl Default for AnsibleConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            proxy: default_ansible_proxy(),
-            proxy_auth: None,
-            proxy_timeout: 30,
-            metadata_ttl: 3600,
-            serve_stale: true,
-        }
-    }
-}
-
-/// NuGet V3 proxy configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct NugetConfig {
-    #[serde(default)]
-    pub enabled: bool,
-    /// Upstream NuGet API (default: https://api.nuget.org)
-    #[serde(default = "default_nuget_proxy")]
-    pub proxy: Option<String>,
-    #[serde(default, skip_serializing)]
-    pub proxy_auth: Option<ProtectedString>,
-    #[serde(default = "default_timeout")]
-    pub proxy_timeout: u64,
-    /// Timeout for metadata requests (registration, version list) in seconds.
-    /// Metadata responses are small JSON; lower timeout improves air-gap stale
-    /// restore by letting the circuit breaker open faster. Default: 2s.
-    #[serde(default = "default_nuget_metadata_timeout")]
-    pub metadata_proxy_timeout: u64,
-    /// Metadata cache TTL in seconds (default: 300 = 5 min).
-    /// -1 = cache forever, 0 = always refetch, >0 = seconds.
-    #[serde(default = "default_metadata_ttl")]
-    pub metadata_ttl: i64,
-    /// Serve stale cached metadata when upstream is unreachable (default: true).
-    #[serde(default = "default_true")]
-    pub serve_stale: bool,
-    /// Upstream NuGet search service URL
-    #[serde(default = "default_nuget_search")]
-    pub search_service: String,
-    /// Upstream NuGet autocomplete service URL
-    #[serde(default = "default_nuget_autocomplete")]
-    pub autocomplete: String,
-}
-
-fn default_nuget_proxy() -> Option<String> {
-    Some("https://api.nuget.org".to_string())
-}
-
-fn default_nuget_search() -> String {
-    "https://azuresearch-usnc.nuget.org/query".to_string()
-}
-
-fn default_nuget_autocomplete() -> String {
-    "https://azuresearch-usnc.nuget.org/autocomplete".to_string()
-}
-
-fn default_nuget_metadata_timeout() -> u64 {
-    2
-}
-
-impl Default for NugetConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            proxy: default_nuget_proxy(),
-            proxy_auth: None,
-            proxy_timeout: 30,
-            metadata_proxy_timeout: 2,
-            metadata_ttl: 300,
-            serve_stale: true,
-            search_service: default_nuget_search(),
-            autocomplete: default_nuget_autocomplete(),
-        }
-    }
-}
-
-/// Dart/Flutter pub registry configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PubDartConfig {
-    #[serde(default)]
-    pub enabled: bool,
-    /// Upstream pub registry (default: https://pub.dev)
-    #[serde(default = "default_pub_proxy")]
-    pub proxy: Option<String>,
-    #[serde(default, skip_serializing)]
-    pub proxy_auth: Option<ProtectedString>,
-    #[serde(default = "default_timeout")]
-    pub proxy_timeout: u64,
-    /// Metadata cache TTL in seconds (default: 300 = 5 min).
-    /// -1 = cache forever, 0 = always refetch, >0 = seconds.
-    #[serde(default = "default_metadata_ttl")]
-    pub metadata_ttl: i64,
-}
-
-fn default_pub_proxy() -> Option<String> {
-    Some("https://pub.dev".to_string())
-}
-
-impl Default for PubDartConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            proxy: default_pub_proxy(),
-            proxy_auth: None,
-            proxy_timeout: 30,
-            metadata_ttl: 300,
-        }
-    }
-}
-
-/// Conan V2 proxy configuration (C/C++ packages)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ConanConfig {
-    #[serde(default)]
-    pub enabled: bool,
-    /// Upstream Conan registry (default: https://center2.conan.io)
-    #[serde(default = "default_conan_proxy")]
-    pub proxy: Option<String>,
-    #[serde(default, skip_serializing)]
-    pub proxy_auth: Option<ProtectedString>,
-    #[serde(default = "default_timeout")]
-    pub proxy_timeout: u64,
-    /// Separate timeout for binary downloads (default: 120s)
-    #[serde(default = "default_go_zip_timeout")]
-    pub proxy_timeout_dl: u64,
-    /// Metadata cache TTL in seconds (default: 300 = 5 min).
-    /// -1 = cache forever, 0 = always refetch, >0 = seconds.
-    #[serde(default = "default_metadata_ttl")]
-    pub metadata_ttl: i64,
-}
-
-fn default_conan_proxy() -> Option<String> {
-    Some("https://center2.conan.io".to_string())
-}
-
-impl Default for ConanConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            proxy: default_conan_proxy(),
-            proxy_auth: None,
-            proxy_timeout: 30,
-            proxy_timeout_dl: 120,
-            metadata_ttl: 300,
-        }
-    }
-}
-
-fn default_docker_timeout() -> u64 {
-    300
-}
-
-fn default_docker_read_timeout() -> u64 {
-    60
-}
-
-fn default_docker_metadata_ttl() -> i64 {
-    -1
-}
-
-fn default_raw_enabled() -> bool {
-    true
-}
-
-fn default_max_file_size() -> u64 {
-    104_857_600 // 100MB
-}
-
-fn default_raw_cache_control() -> String {
-    "no-cache".to_string()
-}
-
-/// CIDR-aware trusted proxy list for X-Forwarded-For validation.
-///
-/// Only connections from trusted proxies have their XFF/X-Real-IP headers
-/// honored. Untrusted sources always use the peer (TCP) IP address.
-#[derive(Debug, Clone)]
-pub struct TrustedProxies {
-    entries: Vec<(std::net::IpAddr, u8)>, // (network address, prefix length)
-}
-
-impl TrustedProxies {
-    /// Parse a comma-separated list of IPs/CIDRs. Invalid entries are skipped with a warning.
-    pub fn parse(input: &str) -> Self {
-        let mut entries = Vec::new();
-        for item in input.split(',') {
-            let item = item.trim();
-            if item.is_empty() {
-                continue;
-            }
-            if let Some((addr_str, prefix_str)) = item.split_once('/') {
-                if let (Ok(addr), Ok(prefix)) = (
-                    addr_str.parse::<std::net::IpAddr>(),
-                    prefix_str.parse::<u8>(),
-                ) {
-                    let max_prefix = if addr.is_ipv4() { 32 } else { 128 };
-                    if prefix <= max_prefix {
-                        if prefix == 0 {
-                            tracing::warn!(
-                                value = %item,
-                                "CIDR /0 matches ALL addresses in this family — all peers will be \
-                                 trusted proxies (X-Forwarded-For honored, IP-based rate limiting disabled)"
-                            );
-                        }
-                        entries.push((addr, prefix));
-                    } else {
-                        tracing::warn!(value = %item, "Invalid CIDR prefix length, skipping");
-                    }
-                } else {
-                    tracing::warn!(value = %item, "Cannot parse CIDR, skipping");
-                }
-            } else if let Ok(addr) = item.parse::<std::net::IpAddr>() {
-                let prefix = if addr.is_ipv4() { 32 } else { 128 };
-                entries.push((addr, prefix));
-            } else {
-                tracing::warn!(value = %item, "Cannot parse IP address, skipping");
-            }
-        }
-        Self { entries }
-    }
-
-    /// Default: loopback only (127.0.0.1 and ::1).
-    pub fn default_loopback() -> Self {
-        Self {
-            entries: vec![
-                (std::net::IpAddr::V4(std::net::Ipv4Addr::LOCALHOST), 32),
-                (std::net::IpAddr::V6(std::net::Ipv6Addr::LOCALHOST), 128),
-            ],
-        }
-    }
-
-    /// Check if an IP address is within the trusted proxy list.
-    pub fn contains(&self, ip: std::net::IpAddr) -> bool {
-        self.entries.iter().any(|(network, prefix)| {
-            match (network, ip) {
-                (std::net::IpAddr::V4(net), std::net::IpAddr::V4(addr)) => {
-                    // /0 matches all addresses in this family (RFC 4632).
-                    // Must check before shift: u32::MAX << 32 overflows.
-                    if *prefix == 0 {
-                        return true;
-                    }
-                    if *prefix >= 32 {
-                        return *net == addr;
-                    }
-                    let net_bits = u32::from(*net);
-                    let addr_bits = u32::from(addr);
-                    let mask = u32::MAX << (32 - prefix);
-                    (net_bits & mask) == (addr_bits & mask)
-                }
-                (std::net::IpAddr::V6(net), std::net::IpAddr::V6(addr)) => {
-                    // /0 matches all addresses in this family (RFC 4632).
-                    // Must check before shift: u128::MAX << 128 overflows.
-                    if *prefix == 0 {
-                        return true;
-                    }
-                    if *prefix >= 128 {
-                        return *net == addr;
-                    }
-                    let net_bits = u128::from(*net);
-                    let addr_bits = u128::from(addr);
-                    let mask = u128::MAX << (128 - prefix);
-                    (net_bits & mask) == (addr_bits & mask)
-                }
-                _ => false, // v4 vs v6 mismatch
-            }
-        })
-    }
-
-    /// Returns true if any entry uses prefix /0 (matches all addresses in its family).
-    pub fn has_prefix_zero(&self) -> bool {
-        self.entries.iter().any(|(_, prefix)| *prefix == 0)
-    }
-}
-
-impl Default for TrustedProxies {
-    fn default() -> Self {
-        Self::default_loopback()
-    }
-}
-
-// TrustedProxies doesn't need serde — it's parsed from a string.
-// Provide a dummy Serialize/Deserialize so AuthConfig can derive them.
-impl Serialize for TrustedProxies {
-    fn serialize<S: serde::Serializer>(
-        &self,
-        serializer: S,
-    ) -> std::result::Result<S::Ok, S::Error> {
-        let parts: Vec<String> = self
-            .entries
-            .iter()
-            .map(|(addr, prefix)| {
-                let max = if addr.is_ipv4() { 32 } else { 128 };
-                if *prefix == max {
-                    addr.to_string()
-                } else {
-                    format!("{}/{}", addr, prefix)
-                }
-            })
-            .collect();
-        parts.join(",").serialize(serializer)
-    }
-}
-
-impl<'de> Deserialize<'de> for TrustedProxies {
-    fn deserialize<D: serde::Deserializer<'de>>(
-        deserializer: D,
-    ) -> std::result::Result<Self, D::Error> {
-        let s = String::deserialize(deserializer)?;
-        Ok(Self::parse(&s))
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AuthConfig {
-    #[serde(default)]
-    pub enabled: bool,
-    /// Allow anonymous read access (pull/download without auth, push requires auth)
-    #[serde(default)]
-    pub anonymous_read: bool,
-    #[serde(default = "default_htpasswd_file")]
-    pub htpasswd_file: String,
-    #[serde(default = "default_token_storage")]
-    pub token_storage: String,
-    /// Trusted proxy IPs/CIDRs — only these sources have XFF/X-Real-IP honored.
-    /// Default: 127.0.0.1,::1 (loopback only).
-    /// ENV: NORA_AUTH_TRUSTED_PROXIES=127.0.0.1,::1,10.0.0.0/8
-    #[serde(default)]
-    pub trusted_proxies: TrustedProxies,
-    /// OIDC providers for workload identity (CI/CD zero-secret auth)
-    #[serde(default)]
-    pub oidc: OidcConfig,
-}
-
-/// OIDC configuration — multiple providers for workload identity auth.
-///
-/// ```toml
-/// [auth.oidc]
-/// enabled = true
-///
-/// [[auth.oidc.providers]]
-/// name = "github-actions"
-/// issuer = "https://token.actions.githubusercontent.com"
-/// audience = "nora"
-/// algorithms = ["RS256", "ES256"]
-/// max_token_lifetime_secs = 900
-/// namespace_scope = ["*"]
-///
-/// [auth.oidc.providers.role_rules]
-/// "repo:myorg/*:ref:refs/heads/main" = "write"
-/// "repo:myorg/*" = "read"
-/// ```
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct OidcConfig {
-    #[serde(default)]
-    pub enabled: bool,
-    /// Clock skew leeway for token validation (seconds)
-    #[serde(default = "default_oidc_leeway")]
-    pub leeway_secs: u64,
-    /// JWKS cache TTL (seconds). Stale keys served on fetch failure.
-    #[serde(default = "default_oidc_jwks_cache_secs")]
-    pub jwks_cache_secs: u64,
-    /// OIDC identity providers
-    #[serde(default)]
-    pub providers: Vec<OidcProvider>,
-}
-
-impl Default for OidcConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            leeway_secs: default_oidc_leeway(),
-            jwks_cache_secs: default_oidc_jwks_cache_secs(),
-            providers: Vec::new(),
-        }
-    }
-}
-
-/// A single OIDC identity provider (e.g., GitHub Actions, GitLab CI).
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct OidcProvider {
-    /// Human-readable name for logs/debugging
-    pub name: String,
-    /// OIDC issuer URL (must match `iss` claim exactly)
-    pub issuer: String,
-    /// Explicit JWKS URI override. If not set, NORA discovers it via
-    /// `{issuer}/.well-known/openid-configuration`. Use when the provider's
-    /// JWKS endpoint doesn't follow the standard `/.well-known/jwks.json` path.
-    #[serde(default)]
-    pub jwks_uri: Option<String>,
-    /// Expected audience (`aud` claim). If empty, audience is not validated.
-    #[serde(default)]
-    pub audience: String,
-    /// Allowed JWT algorithms (default: RS256, ES256). Reject all others.
-    #[serde(default = "default_oidc_algorithms")]
-    pub algorithms: Vec<String>,
-    /// Maximum token lifetime in seconds. Tokens with longer exp-iat are rejected.
-    #[serde(default = "default_oidc_max_lifetime")]
-    pub max_token_lifetime_secs: u64,
-    /// Namespace scope — which NORA namespaces this issuer can access.
-    /// ["*"] = all, ["github/*"] = only repos under github/ prefix.
-    #[serde(default = "default_namespace_scope")]
-    pub namespace_scope: Vec<String>,
-    /// Kill switch — disable this provider without removing config
-    #[serde(default = "default_true")]
-    pub enabled: bool,
-    /// Role rules: glob pattern on `sub` claim → role (read/write/admin).
-    /// First match wins. No match = deny.
-    #[serde(default)]
-    pub role_rules: Vec<OidcRoleRule>,
-}
-
-/// Maps a subject pattern to a NORA role.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct OidcRoleRule {
-    /// Glob pattern matched against the JWT `sub` claim
-    pub pattern: String,
-    /// Role to assign: "read", "write", or "admin"
-    pub role: String,
-}
-
-fn default_oidc_leeway() -> u64 {
-    60
-}
-
-fn default_oidc_jwks_cache_secs() -> u64 {
-    3600
-}
-
-fn default_oidc_algorithms() -> Vec<String> {
-    vec!["RS256".to_string(), "ES256".to_string()]
-}
-
-fn default_oidc_max_lifetime() -> u64 {
-    900 // 15 minutes
-}
-
-fn default_namespace_scope() -> Vec<String> {
-    vec!["*".to_string()]
-}
-
-fn default_htpasswd_file() -> String {
-    "users.htpasswd".to_string()
-}
-
-fn default_token_storage() -> String {
-    "data/tokens".to_string()
-}
-
-fn default_timeout() -> u64 {
-    30
-}
-
-fn default_metadata_ttl() -> i64 {
-    300 // 5 minutes; -1 = cache forever, 0 = always refetch
-}
-
-impl Default for MavenConfig {
-    fn default() -> Self {
-        Self {
-            enabled: true,
-            proxies: vec![MavenProxyEntry::Simple(
-                "https://repo1.maven.org/maven2".to_string(),
-            )],
-            proxy_timeout: 30,
-            checksum_verify: true,
-            immutable_releases: true,
-        }
-    }
-}
-
-impl Default for NpmConfig {
-    fn default() -> Self {
-        Self {
-            enabled: true,
-            proxy: Some("https://registry.npmjs.org".to_string()),
-            proxy_auth: None,
-            proxy_timeout: 30,
-            metadata_ttl: 300,
-        }
-    }
-}
-
-impl Default for PypiConfig {
-    fn default() -> Self {
-        Self {
-            enabled: true,
-            proxy: Some("https://pypi.org/simple/".to_string()),
-            proxy_auth: None,
-            proxy_timeout: 30,
-        }
-    }
-}
-
-impl Default for DockerConfig {
-    fn default() -> Self {
-        Self {
-            enabled: true,
-            proxy_timeout: 300,
-            read_timeout: 60,
-            metadata_ttl: -1,
-            serve_stale: true,
-            upstreams: vec![DockerUpstream {
-                url: "https://registry-1.docker.io".to_string(),
-                auth: None,
-                namespace: None,
-                prefix: None,
-            }],
-        }
-    }
-}
-
-impl Default for RawConfig {
-    fn default() -> Self {
-        Self {
-            enabled: true,
-            max_file_size: 104_857_600, // 100MB
-            cache_control: default_raw_cache_control(),
-        }
-    }
-}
-
-impl Default for AuthConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            anonymous_read: false,
-            htpasswd_file: "users.htpasswd".to_string(),
-            token_storage: "data/tokens".to_string(),
-            trusted_proxies: TrustedProxies::default_loopback(),
-            oidc: OidcConfig::default(),
-        }
-    }
-}
-
-/// Rate limiting configuration
-///
-/// Controls request rate limits for different endpoint types.
-///
-/// # Example
-/// ```toml
-/// [rate_limit]
-/// auth_rps = 1
-/// auth_burst = 5
-/// upload_rps = 200
-/// upload_burst = 500
-/// general_rps = 100
-/// general_burst = 200
-/// ```
-///
-/// # Environment Variables
-/// - `NORA_RATE_LIMIT_AUTH_RPS` - Auth requests per second
-/// - `NORA_RATE_LIMIT_AUTH_BURST` - Auth burst size
-/// - `NORA_RATE_LIMIT_UPLOAD_RPS` - Upload requests per second
-/// - `NORA_RATE_LIMIT_UPLOAD_BURST` - Upload burst size
-/// - `NORA_RATE_LIMIT_GENERAL_RPS` - General requests per second
-/// - `NORA_RATE_LIMIT_GENERAL_BURST` - General burst size
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RateLimitConfig {
-    /// Enable rate limiting (default: true). Set `NORA_RATE_LIMIT_ENABLED=false` to disable.
-    #[serde(default = "default_rate_limit_enabled")]
-    pub enabled: bool,
-    #[serde(default = "default_auth_rps")]
-    pub auth_rps: u64,
-    #[serde(default = "default_auth_burst")]
-    pub auth_burst: u32,
-    #[serde(default = "default_upload_rps")]
-    pub upload_rps: u64,
-    #[serde(default = "default_upload_burst")]
-    pub upload_burst: u32,
-    #[serde(default = "default_general_rps")]
-    pub general_rps: u64,
-    #[serde(default = "default_general_burst")]
-    pub general_burst: u32,
-}
-
-fn default_rate_limit_enabled() -> bool {
-    true
-}
-fn default_auth_rps() -> u64 {
-    1
-}
-fn default_auth_burst() -> u32 {
-    5
-}
-fn default_upload_rps() -> u64 {
-    200
-}
-fn default_upload_burst() -> u32 {
-    500
-}
-fn default_general_rps() -> u64 {
-    100
-}
-fn default_general_burst() -> u32 {
-    200
-}
-
-impl Default for RateLimitConfig {
-    fn default() -> Self {
-        Self {
-            enabled: default_rate_limit_enabled(),
-            auth_rps: default_auth_rps(),
-            auth_burst: default_auth_burst(),
-            upload_rps: default_upload_rps(),
-            upload_burst: default_upload_burst(),
-            general_rps: default_general_rps(),
-            general_burst: default_general_burst(),
-        }
-    }
-}
-
-// ============================================================================
-// GC Configuration
-// ============================================================================
-
-/// Garbage collection configuration.
-///
-/// # Environment Variables
-/// - `NORA_GC_ENABLED` — enable/disable background GC (default: false)
-/// - `NORA_GC_INTERVAL` — interval in seconds between GC runs (default: 86400)
-/// - `NORA_GC_DRY_RUN` — if true, only report orphans without deleting (default: false)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GcConfig {
-    #[serde(default)]
-    pub enabled: bool,
-    #[serde(default = "default_gc_interval")]
-    pub interval: u64,
-    #[serde(default)]
-    pub dry_run: bool,
-}
-
-fn default_gc_interval() -> u64 {
-    86400 // 24 hours
-}
-
-impl Default for GcConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            interval: 86400,
-            dry_run: false,
-        }
-    }
-}
-
-// ============================================================================
-// Retention Configuration
-// ============================================================================
-
-/// A single retention rule applied to a registry.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RetentionRule {
-    /// Registry name (e.g., "docker", "maven", "npm", "pypi", "cargo") or "*" for all
-    pub registry: String,
-    /// Keep the N most recent versions
-    #[serde(default)]
-    pub keep_last: Option<u32>,
-    /// Only delete versions older than N days
-    #[serde(default)]
-    pub older_than_days: Option<u32>,
-    /// Glob patterns that protect versions from deletion
-    #[serde(default)]
-    pub exclude_tags: Vec<String>,
-}
-
-/// Retention policies configuration.
-///
-/// # Environment Variables
-/// - `NORA_RETENTION_ENABLED` — enable/disable background retention (default: false)
-/// - `NORA_RETENTION_INTERVAL` — interval in seconds between runs (default: 86400)
-/// - `NORA_RETENTION_DRY_RUN` — if true, only report what would be deleted (default: false)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RetentionConfig {
-    /// Enable background retention scheduler
-    #[serde(default)]
-    pub enabled: bool,
-    /// Interval in seconds between retention runs (default: 86400 = 24h)
-    #[serde(default = "default_retention_interval")]
-    pub interval: u64,
-    /// If true, only log what would be deleted without actually deleting (default: false)
-    #[serde(default)]
-    pub dry_run: bool,
-    #[serde(default)]
-    pub rules: Vec<RetentionRule>,
-}
-
-fn default_retention_interval() -> u64 {
-    86400 // 24 hours
-}
-
-impl Default for RetentionConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            interval: 86400,
-            dry_run: false,
-            rules: Vec::new(),
-        }
-    }
-}
-
-// ============================================================================
-// Curation Configuration
-// ============================================================================
-
-/// Curation operating mode.
-///
-/// - `off` — curation disabled, all requests pass through (default)
-/// - `audit` — evaluate filters and log decisions, but never block
-/// - `enforce` — evaluate filters and block requests that match a deny rule
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "lowercase")]
-pub enum CurationMode {
-    #[default]
-    Off,
-    Audit,
-    Enforce,
-}
-
-impl std::fmt::Display for CurationMode {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            CurationMode::Off => write!(f, "off"),
-            CurationMode::Audit => write!(f, "audit"),
-            CurationMode::Enforce => write!(f, "enforce"),
-        }
-    }
-}
-
-impl std::str::FromStr for CurationMode {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "off" => Ok(Self::Off),
-            "audit" => Ok(Self::Audit),
-            "enforce" => Ok(Self::Enforce),
-            other => Err(format!(
-                "unknown curation mode {:?} — valid values: off, audit, enforce",
-                other
-            )),
-        }
-    }
-}
-
-/// Behavior when a curation filter returns an error or panics.
-///
-/// - `closed` — treat as blocked (fail-safe, default)
-/// - `open` — treat as allowed (fail-open)
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "lowercase")]
-pub enum CurationOnFailure {
-    #[default]
-    Closed,
-    Open,
-}
-
-/// Curation layer configuration.
-///
-/// # Environment Variables
-/// - `NORA_CURATION_MODE` — off/audit/enforce (default: off)
-/// - `NORA_CURATION_ON_FAILURE` — closed/open (default: closed)
-/// - `NORA_CURATION_ALLOWLIST_PATH` — path to allowlist JSON file
-/// - `NORA_CURATION_BLOCKLIST_PATH` — path to blocklist JSON file
-/// - `NORA_CURATION_BYPASS_TOKEN` — token to bypass curation checks
-/// - `NORA_CURATION_REQUIRE_INTEGRITY` — require integrity metadata (default: false)
-/// - `NORA_CURATION_INTERNAL_NS` — comma-separated glob patterns
-/// - `NORA_CURATION_MIN_RELEASE_AGE` — minimum release age (e.g., "7d", "24h", "1w")
-/// - `NORA_CURATION_QUARANTINE` — quarantine mode: off/observe/enforce (default: off)
-/// - `NORA_CURATION_QUARANTINE_TTL` — quarantine hold duration (e.g., "14d", "24h")
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CurationConfig {
-    #[serde(default)]
-    pub mode: CurationMode,
-    #[serde(default)]
-    pub on_failure: CurationOnFailure,
-    #[serde(default)]
-    pub allowlist_path: Option<String>,
-    #[serde(default)]
-    pub blocklist_path: Option<String>,
-    /// Token to bypass curation. Should only be set via env var, not config file.
-    #[serde(default, skip_serializing)]
-    pub bypass_token: Option<ProtectedString>,
-    #[serde(default)]
-    pub require_integrity: bool,
-    /// Glob patterns for internal namespaces that must never be proxied upstream.
-    /// Always active regardless of curation mode (security boundary).
-    #[serde(default)]
-    pub internal_namespaces: Vec<String>,
-    /// Minimum release age before a package is allowed (e.g., "7d", "24h", "1w").
-    /// Packages published less than this duration ago are blocked.
-    #[serde(default)]
-    pub min_release_age: Option<String>,
-    /// Digest quarantine mode: "off" (default), "observe", or "enforce".
-    /// Tracks first-seen timestamps for proxy-fetched content digests.
-    /// Invalid values are rejected at config parse time (ADR-8: security never silently disabled).
-    #[serde(default)]
-    pub quarantine: Option<QuarantineMode>,
-    /// How long new digests are held in quarantine (e.g., "14d", "24h", "1w").
-    #[serde(default)]
-    pub quarantine_ttl: Option<String>,
-    /// Per-registry curation overrides. Overrides `min_release_age` per registry.
-    #[serde(default)]
-    pub npm: RegistryCurationOverride,
-    #[serde(default)]
-    pub pypi: RegistryCurationOverride,
-    #[serde(default)]
-    pub cargo: RegistryCurationOverride,
-    #[serde(default)]
-    pub go: RegistryCurationOverride,
-    #[serde(default)]
-    pub docker: RegistryCurationOverride,
-    #[serde(default)]
-    pub maven: RegistryCurationOverride,
-    #[serde(default)]
-    pub gems: RegistryCurationOverride,
-    #[serde(default)]
-    pub terraform: RegistryCurationOverride,
-    #[serde(default)]
-    pub ansible: RegistryCurationOverride,
-    #[serde(default)]
-    pub nuget: RegistryCurationOverride,
-    #[serde(rename = "pub", default)]
-    pub pub_dart: RegistryCurationOverride,
-    #[serde(default)]
-    pub conan: RegistryCurationOverride,
-}
-
-/// Per-registry curation override (used within `[curation.{registry}]`).
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct RegistryCurationOverride {
-    /// Override min_release_age for this specific registry.
-    #[serde(default)]
-    pub min_release_age: Option<String>,
-    /// Override quarantine mode for this specific registry.
-    #[serde(default)]
-    pub quarantine: Option<QuarantineMode>,
-    /// Override quarantine TTL for this specific registry.
-    #[serde(default)]
-    pub quarantine_ttl: Option<String>,
-}
-
-impl Default for CurationConfig {
-    fn default() -> Self {
-        Self {
-            mode: CurationMode::Off,
-            on_failure: CurationOnFailure::Closed,
-            allowlist_path: None,
-            blocklist_path: None,
-            bypass_token: None,
-            require_integrity: false,
-            internal_namespaces: Vec::new(),
-            min_release_age: None,
-            quarantine: None,
-            quarantine_ttl: None,
-            npm: RegistryCurationOverride::default(),
-            pypi: RegistryCurationOverride::default(),
-            cargo: RegistryCurationOverride::default(),
-            go: RegistryCurationOverride::default(),
-            docker: RegistryCurationOverride::default(),
-            maven: RegistryCurationOverride::default(),
-            gems: RegistryCurationOverride::default(),
-            terraform: RegistryCurationOverride::default(),
-            ansible: RegistryCurationOverride::default(),
-            nuget: RegistryCurationOverride::default(),
-            pub_dart: RegistryCurationOverride::default(),
-            conan: RegistryCurationOverride::default(),
-        }
-    }
-}
-
-// ============================================================================
-// Circuit Breaker
-// ============================================================================
-
-fn default_cb_enabled() -> bool {
-    false
-}
-fn default_cb_threshold() -> u32 {
-    5
-}
-fn default_cb_reset_timeout() -> u64 {
-    30
-}
-
-/// Upstream proxy circuit breaker configuration.
-///
-/// Experimental — disabled by default. When enabled, tracks per-registry
-/// upstream failures and fails fast (503) when a registry is known to be down.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CircuitBreakerConfig {
-    /// Enable circuit breaker (default: false)
-    #[serde(default = "default_cb_enabled")]
-    pub enabled: bool,
-    /// Number of consecutive failures before opening the circuit (default: 5)
-    #[serde(default = "default_cb_threshold")]
-    pub failure_threshold: u32,
-    /// Seconds to wait before probing a failed upstream (default: 30)
-    #[serde(default = "default_cb_reset_timeout")]
-    pub reset_timeout: u64,
-    /// Per-registry overrides keyed by circuit breaker key (e.g. "docker:https://registry-1.docker.io").
-    #[serde(default)]
-    pub overrides: HashMap<String, CircuitBreakerOverride>,
-}
-
-/// Per-registry circuit breaker threshold overrides.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CircuitBreakerOverride {
-    pub failure_threshold: Option<u32>,
-    pub reset_timeout: Option<u64>,
-}
-
-impl Default for CircuitBreakerConfig {
-    fn default() -> Self {
-        Self {
-            enabled: default_cb_enabled(),
-            failure_threshold: default_cb_threshold(),
-            reset_timeout: default_cb_reset_timeout(),
-            overrides: HashMap::new(),
-        }
-    }
-}
-
-// ============================================================================
-// Registries Section (nginx-style enable)
-// ============================================================================
-
-/// Top-level `[registries]` section for declarative registry selection.
-///
-/// # Example
-/// ```toml
-/// [registries]
-/// enable = ["docker", "npm", "pypi"]
-///
-/// # Or enable all except some:
-/// # enable = ["all", "-maven"]
-///
-/// # Or enable everything:
-/// # enable = "all"
-/// ```
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct RegistriesSection {
-    #[serde(default)]
-    pub enable: Option<EnableSpec>,
-}
-
-/// What registries to enable — a single string or list of strings.
-///
-/// Supports:
-/// - `"all"` — all 13 registries
-/// - `"docker"` — single registry
-/// - `["docker", "npm", "pypi"]` — explicit list
-/// - `["all", "-maven"]` — all except maven
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(untagged)]
-pub enum EnableSpec {
-    /// Single string: `enable = "all"` or `enable = "docker"`
-    Single(String),
-    /// List of strings: `enable = ["docker", "npm"]`
-    List(Vec<String>),
-}
-
-impl EnableSpec {
-    /// Parse from comma-separated env var string.
-    /// E.g. `"docker,npm,pypi"` or `"all,-maven"`.
-    pub fn from_env_str(s: &str) -> Self {
-        let items: Vec<String> = s
-            .split(',')
-            .map(|item| item.trim().to_lowercase())
-            .filter(|item| !item.is_empty())
-            .collect();
-        match items.len() {
-            1 => {
-                // Safety: we just checked len() == 1
-                let item = items.into_iter().next().unwrap_or_default();
-                EnableSpec::Single(item)
-            }
-            _ => EnableSpec::List(items),
-        }
-    }
-
-    /// Resolve the spec into a concrete set of RegistryTypes.
-    ///
-    /// Rules:
-    /// - `"all"` → all 13 registries
-    /// - `"-name"` → exclusion (only valid when `"all"` is present)
-    /// - `"name"` → inclusion
-    /// - Unknown name → Err
-    /// - Empty result → Err
-    pub fn resolve(&self) -> Result<HashSet<RegistryType>, String> {
-        let items = match self {
-            EnableSpec::Single(s) => vec![s.clone()],
-            EnableSpec::List(v) => v.clone(),
-        };
-
-        if items.is_empty() {
-            return Err("registries.enable must not be empty".to_string());
-        }
-
-        let has_all = items.iter().any(|s| s == "all");
-        let exclusions: Vec<&str> = items
-            .iter()
-            .filter(|s| s.starts_with('-'))
-            .map(|s| s.as_str())
-            .collect();
-        let inclusions: Vec<&str> = items
-            .iter()
-            .filter(|s| *s != "all" && !s.starts_with('-'))
-            .map(|s| s.as_str())
-            .collect();
-
-        // Exclusions without "all" is an error
-        if !exclusions.is_empty() && !has_all {
-            return Err(format!(
-                "exclusion entries ({}) require \"all\" in the list",
-                exclusions.join(", ")
-            ));
-        }
-
-        // "all" with inclusions is ambiguous
-        if has_all && !inclusions.is_empty() {
-            return Err(format!(
-                "\"all\" cannot be combined with inclusions ({}); use \"all\" with exclusions like \"-maven\"",
-                inclusions.join(", ")
-            ));
-        }
-
-        if has_all {
-            // Start with all, then remove exclusions
-            let mut set: HashSet<RegistryType> = RegistryType::all().iter().copied().collect();
-            for ex in &exclusions {
-                let name = &ex[1..]; // strip leading '-'
-                match RegistryType::from_str_opt(name) {
-                    Some(rt) => {
-                        set.remove(&rt);
-                    }
-                    None => {
-                        return Err(format!("unknown registry in exclusion: \"{}\"", ex));
-                    }
-                }
-            }
-            if set.is_empty() {
-                return Err("all registries excluded — at least one must be enabled".to_string());
-            }
-            Ok(set)
-        } else {
-            // Explicit inclusion list
-            let mut set = HashSet::new();
-            for name in &inclusions {
-                match RegistryType::from_str_opt(name) {
-                    Some(rt) => {
-                        set.insert(rt);
-                    }
-                    None => {
-                        return Err(format!("unknown registry: \"{}\"", name));
-                    }
-                }
-            }
-            if set.is_empty() {
-                return Err("registries.enable must not be empty".to_string());
-            }
-            Ok(set)
-        }
-    }
-}
-
-/// Audit log configuration.
-///
-/// Controls where audit events are written:
-/// - `file`   — write to {storage_path}/audit.jsonl (default)
-/// - `stdout` — write JSONL to stderr (12-factor compatible)
-/// - `both`   — write to file AND stderr
-/// - `off`    — disable audit logging
-///
-/// ENV: `NORA_AUDIT_LOG=file|stdout|both|off`
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AuditConfig {
-    #[serde(default)]
-    pub mode: crate::audit::AuditMode,
-}
-
-impl Default for AuditConfig {
-    fn default() -> Self {
-        Self {
-            mode: crate::audit::AuditMode::File,
-        }
-    }
-}
-
-/// Parse an env var value into `target`, warning on failure (#537).
-///
-/// If `val` parses successfully, `*target` is updated. If parsing fails,
-/// a `warn!` is emitted with the variable name and invalid value, and
-/// `*target` is left unchanged (keeps the TOML/default value).
-fn parse_env_warn<T: std::str::FromStr + std::fmt::Display>(name: &str, val: &str, target: &mut T) {
-    match val.parse::<T>() {
-        Ok(parsed) => {
-            *target = parsed;
-        }
-        Err(_) => {
-            tracing::warn!(
-                var = name,
-                value = val,
-                "env override ignored: failed to parse value"
-            );
-        }
-    }
 }
 
 impl Config {
@@ -2151,588 +709,37 @@ impl Config {
     /// audit mode, storage mode) has an unrecognized value — typos must not silently
     /// disable security or misconfigure storage.
     fn apply_env_overrides(&mut self) -> Result<(), String> {
-        // Server config
-        if let Ok(val) = env::var("NORA_HOST") {
-            self.server.host = val;
-        }
-        if let Ok(val) = env::var("NORA_PORT") {
-            parse_env_warn("NORA_PORT", &val, &mut self.server.port);
-        }
-        if let Ok(val) = env::var("NORA_PUBLIC_URL") {
-            self.server.public_url = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_BODY_LIMIT_MB") {
-            parse_env_warn("NORA_BODY_LIMIT_MB", &val, &mut self.server.body_limit_mb);
-        }
+        // Server + TLS
+        self.server.apply_env_overrides();
+        self.tls.apply_env_overrides();
 
-        // TLS config
-        if let Ok(val) = env::var("NORA_TLS_CA_CERT") {
-            self.tls.ca_cert = if val.is_empty() { None } else { Some(val) };
-        }
+        // Storage (fail-closed: unknown NORA_STORAGE_MODE is fatal)
+        self.storage.apply_env_overrides()?;
 
-        // Storage config
-        if let Ok(val) = env::var("NORA_STORAGE_MODE") {
-            self.storage.mode = match val.to_lowercase().as_str() {
-                "local" | "filesystem" => StorageMode::Local,
-                "s3" => StorageMode::S3,
-                other => {
-                    return Err(format!(
-                        "NORA_STORAGE_MODE={:?} is invalid — valid values: local, s3",
-                        other
-                    ))
-                }
-            };
-        }
-        if let Ok(val) = env::var("NORA_STORAGE_PATH") {
-            self.storage.path = val;
-        }
-        if let Ok(val) = env::var("NORA_STORAGE_S3_URL") {
-            self.storage.s3_url = val;
-        }
-        if let Ok(val) = env::var("NORA_STORAGE_BUCKET") {
-            self.storage.bucket = val;
-        }
-        if let Ok(val) = env::var("NORA_STORAGE_S3_ACCESS_KEY") {
-            self.storage.s3_access_key = if val.is_empty() {
-                None
-            } else {
-                Some(ProtectedString::new(val))
-            };
-        }
-        if let Ok(val) = env::var("NORA_STORAGE_S3_SECRET_KEY") {
-            self.storage.s3_secret_key = if val.is_empty() {
-                None
-            } else {
-                Some(ProtectedString::new(val))
-            };
-        }
-        if let Ok(val) = env::var("NORA_STORAGE_S3_REGION") {
-            self.storage.s3_region = val;
-        }
+        // Auth
+        self.auth.apply_env_overrides();
 
-        // Auth config
-        if let Ok(val) = env::var("NORA_AUTH_ENABLED") {
-            self.auth.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_AUTH_ANONYMOUS_READ") {
-            self.auth.anonymous_read = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_AUTH_HTPASSWD_FILE") {
-            self.auth.htpasswd_file = val;
-        }
-        if let Ok(val) = env::var("NORA_AUTH_TRUSTED_PROXIES") {
-            self.auth.trusted_proxies = TrustedProxies::parse(&val);
-        }
-        if let Ok(val) = env::var("NORA_AUTH_OIDC_ENABLED") {
-            self.auth.oidc.enabled = val.to_lowercase() == "true" || val == "1";
-        }
+        // Registry configs (each handles its own NORA_*_ENABLED + format-specific vars)
+        self.docker.apply_env_overrides();
+        self.maven.apply_env_overrides();
+        self.npm.apply_env_overrides();
+        self.pypi.apply_env_overrides();
+        self.go.apply_env_overrides();
+        self.cargo.apply_env_overrides();
+        self.raw.apply_env_overrides();
+        self.gems.apply_env_overrides();
+        self.terraform.apply_env_overrides();
+        self.ansible.apply_env_overrides();
+        self.nuget.apply_env_overrides();
+        self.pub_dart.apply_env_overrides();
+        self.conan.apply_env_overrides();
 
-        // Registry enabled flags
-        if let Ok(val) = env::var("NORA_DOCKER_ENABLED") {
-            self.docker.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_MAVEN_ENABLED") {
-            self.maven.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_NPM_ENABLED") {
-            self.npm.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_CARGO_ENABLED") {
-            self.cargo.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_PYPI_ENABLED") {
-            self.pypi.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_GO_ENABLED") {
-            self.go.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_GEMS_ENABLED") {
-            self.gems.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_TF_ENABLED") {
-            self.terraform.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_ANSIBLE_ENABLED") {
-            self.ansible.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_NUGET_ENABLED") {
-            self.nuget.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_PUB_ENABLED") {
-            self.pub_dart.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_CONAN_ENABLED") {
-            self.conan.enabled = val.to_lowercase() == "true" || val == "1";
-        }
+        // Rate limit, GC, retention
+        self.rate_limit.apply_env_overrides();
+        self.gc.apply_env_overrides();
+        self.retention.apply_env_overrides();
 
-        // Maven config — supports "url1,url2" or "url1|auth1,url2|auth2"
-        if let Ok(val) = env::var("NORA_MAVEN_PROXIES") {
-            self.maven.proxies = val
-                .split(',')
-                .filter(|s| !s.is_empty())
-                .map(|s| {
-                    let parts: Vec<&str> = s.trim().splitn(2, '|').collect();
-                    if parts.len() > 1 {
-                        MavenProxyEntry::Full(MavenProxy {
-                            url: parts[0].to_string(),
-                            auth: Some(ProtectedString::from(parts[1])),
-                        })
-                    } else {
-                        MavenProxyEntry::Simple(parts[0].to_string())
-                    }
-                })
-                .collect();
-        }
-        if let Ok(val) = env::var("NORA_MAVEN_PROXY_TIMEOUT") {
-            parse_env_warn(
-                "NORA_MAVEN_PROXY_TIMEOUT",
-                &val,
-                &mut self.maven.proxy_timeout,
-            );
-        }
-        if let Ok(val) = env::var("NORA_MAVEN_CHECKSUM_VERIFY") {
-            self.maven.checksum_verify = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_MAVEN_IMMUTABLE_RELEASES") {
-            self.maven.immutable_releases = val.to_lowercase() == "true" || val == "1";
-        }
-
-        // npm config
-        if let Ok(val) = env::var("NORA_NPM_PROXY") {
-            self.npm.proxy = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_NPM_PROXY_TIMEOUT") {
-            parse_env_warn("NORA_NPM_PROXY_TIMEOUT", &val, &mut self.npm.proxy_timeout);
-        }
-        if let Ok(val) = env::var("NORA_NPM_METADATA_TTL") {
-            parse_env_warn("NORA_NPM_METADATA_TTL", &val, &mut self.npm.metadata_ttl);
-        }
-
-        // npm proxy auth
-        if let Ok(val) = env::var("NORA_NPM_PROXY_AUTH") {
-            self.npm.proxy_auth = if val.is_empty() {
-                None
-            } else {
-                Some(ProtectedString::new(val))
-            };
-        }
-
-        // PyPI config
-        if let Ok(val) = env::var("NORA_PYPI_PROXY") {
-            self.pypi.proxy = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_PYPI_PROXY_TIMEOUT") {
-            parse_env_warn(
-                "NORA_PYPI_PROXY_TIMEOUT",
-                &val,
-                &mut self.pypi.proxy_timeout,
-            );
-        }
-
-        // PyPI proxy auth
-        if let Ok(val) = env::var("NORA_PYPI_PROXY_AUTH") {
-            self.pypi.proxy_auth = if val.is_empty() {
-                None
-            } else {
-                Some(ProtectedString::new(val))
-            };
-        }
-
-        // Docker config
-        if let Ok(val) = env::var("NORA_DOCKER_PROXY_TIMEOUT") {
-            parse_env_warn(
-                "NORA_DOCKER_PROXY_TIMEOUT",
-                &val,
-                &mut self.docker.proxy_timeout,
-            );
-        }
-        if let Ok(val) = env::var("NORA_DOCKER_READ_TIMEOUT") {
-            parse_env_warn(
-                "NORA_DOCKER_READ_TIMEOUT",
-                &val,
-                &mut self.docker.read_timeout,
-            );
-        }
-        if let Ok(val) = env::var("NORA_DOCKER_METADATA_TTL") {
-            parse_env_warn(
-                "NORA_DOCKER_METADATA_TTL",
-                &val,
-                &mut self.docker.metadata_ttl,
-            );
-        }
-        if let Ok(val) = env::var("NORA_DOCKER_SERVE_STALE") {
-            self.docker.serve_stale = !matches!(val.as_str(), "false" | "0");
-        }
-        // NORA_DOCKER_PROXIES format: "url1,url2" or "url1|auth1,url2|auth2"
-        // or "url1|auth1|prefix1,url2||prefix2" (empty auth with prefix)
-        // Backward compat: NORA_DOCKER_UPSTREAMS still works but is deprecated
-        if let Ok(val) =
-            env::var("NORA_DOCKER_PROXIES").or_else(|_| env::var("NORA_DOCKER_UPSTREAMS"))
-        {
-            if env::var("NORA_DOCKER_PROXIES").is_err() {
-                tracing::warn!("NORA_DOCKER_UPSTREAMS is deprecated, use NORA_DOCKER_PROXIES");
-            }
-            self.docker.upstreams = val
-                .split(',')
-                .filter(|s| !s.is_empty())
-                .map(|s| {
-                    let parts: Vec<&str> = s.trim().splitn(3, '|').collect();
-                    let auth = parts.get(1).and_then(|a| {
-                        if a.is_empty() {
-                            None
-                        } else {
-                            Some(ProtectedString::from(*a))
-                        }
-                    });
-                    let prefix = parts.get(2).and_then(|p| {
-                        if p.is_empty() {
-                            None
-                        } else {
-                            Some(p.to_string())
-                        }
-                    });
-                    DockerUpstream {
-                        url: parts[0].to_string(),
-                        auth,
-                        namespace: None,
-                        prefix,
-                    }
-                })
-                .collect();
-            if self.docker.upstreams.iter().any(|u| u.auth.is_some()) {
-                tracing::warn!(
-                    "Docker upstream credentials passed via NORA_DOCKER_PROXIES environment variable. \
-                     For production use config.toml with [[docker.upstreams]] and mount credentials from a Kubernetes Secret."
-                );
-            }
-        }
-
-        // Go config
-        if let Ok(val) = env::var("NORA_GO_PROXY") {
-            self.go.proxy = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_GO_PROXY_AUTH") {
-            self.go.proxy_auth = if val.is_empty() {
-                None
-            } else {
-                Some(ProtectedString::new(val))
-            };
-        }
-        if let Ok(val) = env::var("NORA_GO_PROXY_TIMEOUT") {
-            parse_env_warn("NORA_GO_PROXY_TIMEOUT", &val, &mut self.go.proxy_timeout);
-        }
-        if let Ok(val) = env::var("NORA_GO_PROXY_TIMEOUT_ZIP") {
-            parse_env_warn(
-                "NORA_GO_PROXY_TIMEOUT_ZIP",
-                &val,
-                &mut self.go.proxy_timeout_zip,
-            );
-        }
-        if let Ok(val) = env::var("NORA_GO_MAX_ZIP_SIZE") {
-            parse_env_warn("NORA_GO_MAX_ZIP_SIZE", &val, &mut self.go.max_zip_size);
-        }
-
-        // Cargo config
-        if let Ok(val) = env::var("NORA_CARGO_PROXY") {
-            self.cargo.proxy = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_CARGO_PROXY_TIMEOUT") {
-            parse_env_warn(
-                "NORA_CARGO_PROXY_TIMEOUT",
-                &val,
-                &mut self.cargo.proxy_timeout,
-            );
-        }
-        if let Ok(val) = env::var("NORA_CARGO_PROXY_AUTH") {
-            self.cargo.proxy_auth = if val.is_empty() {
-                None
-            } else {
-                Some(ProtectedString::new(val))
-            };
-        }
-
-        // Raw config
-        if let Ok(val) = env::var("NORA_RAW_ENABLED") {
-            self.raw.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_RAW_MAX_FILE_SIZE") {
-            parse_env_warn("NORA_RAW_MAX_FILE_SIZE", &val, &mut self.raw.max_file_size);
-        }
-        if let Ok(val) = env::var("NORA_RAW_CACHE_CONTROL") {
-            self.raw.cache_control = val;
-        }
-
-        // Gems config
-        if let Ok(val) = env::var("NORA_GEMS_PROXY") {
-            self.gems.proxy = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_GEMS_PROXY_AUTH") {
-            self.gems.proxy_auth = if val.is_empty() {
-                None
-            } else {
-                Some(ProtectedString::new(val))
-            };
-        }
-        if let Ok(val) = env::var("NORA_GEMS_PROXY_TIMEOUT") {
-            parse_env_warn(
-                "NORA_GEMS_PROXY_TIMEOUT",
-                &val,
-                &mut self.gems.proxy_timeout,
-            );
-        }
-        if let Ok(val) = env::var("NORA_GEMS_METADATA_TTL") {
-            parse_env_warn("NORA_GEMS_METADATA_TTL", &val, &mut self.gems.metadata_ttl);
-        }
-
-        // Terraform config
-        if let Ok(val) = env::var("NORA_TF_PROXY") {
-            self.terraform.proxy = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_TF_PROXY_AUTH") {
-            self.terraform.proxy_auth = if val.is_empty() {
-                None
-            } else {
-                Some(ProtectedString::new(val))
-            };
-        }
-        if let Ok(val) = env::var("NORA_TF_PROXY_TIMEOUT") {
-            parse_env_warn(
-                "NORA_TF_PROXY_TIMEOUT",
-                &val,
-                &mut self.terraform.proxy_timeout,
-            );
-        }
-        if let Ok(val) = env::var("NORA_TF_PROXY_TIMEOUT_DL") {
-            parse_env_warn(
-                "NORA_TF_PROXY_TIMEOUT_DL",
-                &val,
-                &mut self.terraform.proxy_timeout_dl,
-            );
-        }
-        if let Ok(val) = env::var("NORA_TF_METADATA_TTL") {
-            parse_env_warn(
-                "NORA_TF_METADATA_TTL",
-                &val,
-                &mut self.terraform.metadata_ttl,
-            );
-        }
-        if let Ok(val) = env::var("NORA_TF_SERVE_STALE") {
-            self.terraform.serve_stale = !matches!(val.as_str(), "false" | "0");
-        }
-
-        // Ansible Galaxy config
-        if let Ok(val) = env::var("NORA_ANSIBLE_PROXY") {
-            self.ansible.proxy = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_ANSIBLE_PROXY_AUTH") {
-            self.ansible.proxy_auth = if val.is_empty() {
-                None
-            } else {
-                Some(ProtectedString::new(val))
-            };
-        }
-        if let Ok(val) = env::var("NORA_ANSIBLE_PROXY_TIMEOUT") {
-            parse_env_warn(
-                "NORA_ANSIBLE_PROXY_TIMEOUT",
-                &val,
-                &mut self.ansible.proxy_timeout,
-            );
-        }
-        if let Ok(val) = env::var("NORA_ANSIBLE_METADATA_TTL") {
-            parse_env_warn(
-                "NORA_ANSIBLE_METADATA_TTL",
-                &val,
-                &mut self.ansible.metadata_ttl,
-            );
-        }
-        if let Ok(val) = env::var("NORA_ANSIBLE_SERVE_STALE") {
-            self.ansible.serve_stale = !matches!(val.as_str(), "false" | "0");
-        }
-
-        // NuGet config
-        if let Ok(val) = env::var("NORA_NUGET_PROXY") {
-            self.nuget.proxy = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_NUGET_PROXY_AUTH") {
-            self.nuget.proxy_auth = if val.is_empty() {
-                None
-            } else {
-                Some(ProtectedString::new(val))
-            };
-        }
-        if let Ok(val) = env::var("NORA_NUGET_PROXY_TIMEOUT") {
-            parse_env_warn(
-                "NORA_NUGET_PROXY_TIMEOUT",
-                &val,
-                &mut self.nuget.proxy_timeout,
-            );
-        }
-        if let Ok(val) = env::var("NORA_NUGET_METADATA_TIMEOUT") {
-            parse_env_warn(
-                "NORA_NUGET_METADATA_TIMEOUT",
-                &val,
-                &mut self.nuget.metadata_proxy_timeout,
-            );
-        }
-        if let Ok(val) = env::var("NORA_NUGET_METADATA_TTL") {
-            parse_env_warn(
-                "NORA_NUGET_METADATA_TTL",
-                &val,
-                &mut self.nuget.metadata_ttl,
-            );
-        }
-        if let Ok(val) = env::var("NORA_NUGET_SERVE_STALE") {
-            self.nuget.serve_stale = !matches!(val.as_str(), "false" | "0");
-        }
-        if let Ok(val) = env::var("NORA_NUGET_SEARCH_SERVICE") {
-            if !val.is_empty() {
-                self.nuget.search_service = val;
-            }
-        }
-        if let Ok(val) = env::var("NORA_NUGET_AUTOCOMPLETE") {
-            if !val.is_empty() {
-                self.nuget.autocomplete = val;
-            }
-        }
-
-        // pub.dev config
-        if let Ok(val) = env::var("NORA_PUB_PROXY") {
-            self.pub_dart.proxy = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_PUB_PROXY_AUTH") {
-            self.pub_dart.proxy_auth = if val.is_empty() {
-                None
-            } else {
-                Some(ProtectedString::new(val))
-            };
-        }
-        if let Ok(val) = env::var("NORA_PUB_PROXY_TIMEOUT") {
-            parse_env_warn(
-                "NORA_PUB_PROXY_TIMEOUT",
-                &val,
-                &mut self.pub_dart.proxy_timeout,
-            );
-        }
-        if let Ok(val) = env::var("NORA_PUB_METADATA_TTL") {
-            parse_env_warn(
-                "NORA_PUB_METADATA_TTL",
-                &val,
-                &mut self.pub_dart.metadata_ttl,
-            );
-        }
-
-        // Conan proxy config
-        if let Ok(val) = env::var("NORA_CONAN_PROXY") {
-            self.conan.proxy = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_CONAN_PROXY_AUTH") {
-            self.conan.proxy_auth = if val.is_empty() {
-                None
-            } else {
-                Some(ProtectedString::new(val))
-            };
-        }
-        if let Ok(val) = env::var("NORA_CONAN_PROXY_TIMEOUT") {
-            parse_env_warn(
-                "NORA_CONAN_PROXY_TIMEOUT",
-                &val,
-                &mut self.conan.proxy_timeout,
-            );
-        }
-        if let Ok(val) = env::var("NORA_CONAN_PROXY_TIMEOUT_DL") {
-            parse_env_warn(
-                "NORA_CONAN_PROXY_TIMEOUT_DL",
-                &val,
-                &mut self.conan.proxy_timeout_dl,
-            );
-        }
-        if let Ok(val) = env::var("NORA_CONAN_METADATA_TTL") {
-            parse_env_warn(
-                "NORA_CONAN_METADATA_TTL",
-                &val,
-                &mut self.conan.metadata_ttl,
-            );
-        }
-
-        // Token storage
-        if let Ok(val) = env::var("NORA_AUTH_TOKEN_STORAGE") {
-            self.auth.token_storage = val;
-        }
-
-        // Rate limit config
-        if let Ok(val) = env::var("NORA_RATE_LIMIT_ENABLED") {
-            self.rate_limit.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_RATE_LIMIT_AUTH_RPS") {
-            parse_env_warn(
-                "NORA_RATE_LIMIT_AUTH_RPS",
-                &val,
-                &mut self.rate_limit.auth_rps,
-            );
-        }
-        if let Ok(val) = env::var("NORA_RATE_LIMIT_AUTH_BURST") {
-            parse_env_warn(
-                "NORA_RATE_LIMIT_AUTH_BURST",
-                &val,
-                &mut self.rate_limit.auth_burst,
-            );
-        }
-        if let Ok(val) = env::var("NORA_RATE_LIMIT_UPLOAD_RPS") {
-            parse_env_warn(
-                "NORA_RATE_LIMIT_UPLOAD_RPS",
-                &val,
-                &mut self.rate_limit.upload_rps,
-            );
-        }
-        if let Ok(val) = env::var("NORA_RATE_LIMIT_UPLOAD_BURST") {
-            parse_env_warn(
-                "NORA_RATE_LIMIT_UPLOAD_BURST",
-                &val,
-                &mut self.rate_limit.upload_burst,
-            );
-        }
-        if let Ok(val) = env::var("NORA_RATE_LIMIT_GENERAL_RPS") {
-            parse_env_warn(
-                "NORA_RATE_LIMIT_GENERAL_RPS",
-                &val,
-                &mut self.rate_limit.general_rps,
-            );
-        }
-        if let Ok(val) = env::var("NORA_RATE_LIMIT_GENERAL_BURST") {
-            parse_env_warn(
-                "NORA_RATE_LIMIT_GENERAL_BURST",
-                &val,
-                &mut self.rate_limit.general_burst,
-            );
-        }
-
-        // GC config
-        if let Ok(val) = env::var("NORA_GC_ENABLED") {
-            self.gc.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_GC_INTERVAL") {
-            parse_env_warn("NORA_GC_INTERVAL", &val, &mut self.gc.interval);
-        }
-        if let Ok(val) = env::var("NORA_GC_DRY_RUN") {
-            self.gc.dry_run = val.to_lowercase() == "true" || val == "1";
-        }
-
-        // Retention scheduler config
-        if let Ok(val) = env::var("NORA_RETENTION_ENABLED") {
-            self.retention.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_RETENTION_INTERVAL") {
-            parse_env_warn(
-                "NORA_RETENTION_INTERVAL",
-                &val,
-                &mut self.retention.interval,
-            );
-        }
-        if let Ok(val) = env::var("NORA_RETENTION_DRY_RUN") {
-            self.retention.dry_run = val.to_lowercase() == "true" || val == "1";
-        }
-
-        // Secrets config
+        // Secrets — SecretsConfig lives in crate::secrets, no apply_env_overrides method
         if let Ok(val) = env::var("NORA_SECRETS_PROVIDER") {
             self.secrets.provider = val;
         }
@@ -2740,115 +747,10 @@ impl Config {
             self.secrets.clear_env = val.to_lowercase() == "true" || val == "1";
         }
 
-        // Curation config
-        if let Ok(val) = env::var("NORA_CURATION_MODE") {
-            self.curation.mode = val
-                .parse::<CurationMode>()
-                .map_err(|e| format!("NORA_CURATION_MODE={:?}: {}", val, e))?;
-        }
-        if let Ok(val) = env::var("NORA_CURATION_ON_FAILURE") {
-            self.curation.on_failure = match val.to_lowercase().as_str() {
-                "open" => CurationOnFailure::Open,
-                _ => CurationOnFailure::Closed,
-            };
-        }
-        if let Ok(val) = env::var("NORA_CURATION_ALLOWLIST_PATH") {
-            self.curation.allowlist_path = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_CURATION_BLOCKLIST_PATH") {
-            self.curation.blocklist_path = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_CURATION_BYPASS_TOKEN") {
-            self.curation.bypass_token = if val.is_empty() {
-                None
-            } else {
-                Some(ProtectedString::new(val))
-            };
-        }
-        if let Ok(val) = env::var("NORA_CURATION_REQUIRE_INTEGRITY") {
-            self.curation.require_integrity = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_CURATION_INTERNAL_NS") {
-            self.curation.internal_namespaces = if val.is_empty() {
-                Vec::new()
-            } else {
-                val.split(',').map(|s| s.trim().to_string()).collect()
-            };
-        }
-        if let Ok(val) = env::var("NORA_CURATION_MIN_RELEASE_AGE") {
-            self.curation.min_release_age = if val.is_empty() { None } else { Some(val) };
-        }
-        if let Ok(val) = env::var("NORA_CURATION_QUARANTINE") {
-            self.curation.quarantine = if val.is_empty() {
-                None
-            } else {
-                Some(
-                    val.parse::<QuarantineMode>()
-                        .map_err(|e| format!("NORA_CURATION_QUARANTINE={:?}: {}", val, e))?,
-                )
-            };
-        }
-        if let Ok(val) = env::var("NORA_CURATION_QUARANTINE_TTL") {
-            self.curation.quarantine_ttl = if val.is_empty() { None } else { Some(val) };
-        }
-
-        // Circuit breaker config
-        if let Ok(val) = env::var("NORA_CB_ENABLED") {
-            self.circuit_breaker.enabled = val.to_lowercase() == "true" || val == "1";
-        }
-        if let Ok(val) = env::var("NORA_CB_THRESHOLD") {
-            parse_env_warn(
-                "NORA_CB_THRESHOLD",
-                &val,
-                &mut self.circuit_breaker.failure_threshold,
-            );
-        }
-        if let Ok(val) = env::var("NORA_CB_RESET_TIMEOUT") {
-            parse_env_warn(
-                "NORA_CB_RESET_TIMEOUT",
-                &val,
-                &mut self.circuit_breaker.reset_timeout,
-            );
-        }
-
-        // Per-registry curation overrides
-        for (env_suffix, field) in [
-            ("NPM", &mut self.curation.npm),
-            ("PYPI", &mut self.curation.pypi),
-            ("CARGO", &mut self.curation.cargo),
-            ("GO", &mut self.curation.go),
-            ("DOCKER", &mut self.curation.docker),
-            ("MAVEN", &mut self.curation.maven),
-            ("GEMS", &mut self.curation.gems),
-            ("TERRAFORM", &mut self.curation.terraform),
-            ("ANSIBLE", &mut self.curation.ansible),
-            ("NUGET", &mut self.curation.nuget),
-            ("PUB", &mut self.curation.pub_dart),
-            ("CONAN", &mut self.curation.conan),
-        ] {
-            if let Ok(val) = env::var(format!("NORA_CURATION_{}_MIN_RELEASE_AGE", env_suffix)) {
-                field.min_release_age = if val.is_empty() { None } else { Some(val) };
-            }
-            if let Ok(val) = env::var(format!("NORA_CURATION_{}_QUARANTINE", env_suffix)) {
-                field.quarantine = if val.is_empty() {
-                    None
-                } else {
-                    Some(val.parse::<QuarantineMode>().map_err(|e| {
-                        format!("NORA_CURATION_{}_QUARANTINE={:?}: {}", env_suffix, val, e)
-                    })?)
-                };
-            }
-            if let Ok(val) = env::var(format!("NORA_CURATION_{}_QUARANTINE_TTL", env_suffix)) {
-                field.quarantine_ttl = if val.is_empty() { None } else { Some(val) };
-            }
-        }
-
-        // Audit config
-        if let Ok(val) = env::var("NORA_AUDIT_LOG") {
-            self.audit.mode = val
-                .parse::<crate::audit::AuditMode>()
-                .map_err(|e| format!("NORA_AUDIT_LOG={:?}: {}", val, e))?;
-        }
+        // Security-critical: parse errors are fatal (fail-fatal security enums)
+        self.curation.apply_env_overrides()?;
+        self.circuit_breaker.apply_env_overrides();
+        self.audit.apply_env_overrides()?;
 
         Ok(())
     }
@@ -2903,6 +805,8 @@ impl Default for Config {
 #[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;
+    use crate::digest_quarantine::QuarantineMode;
+    use crate::secrets::ProtectedString;
     use std::sync::{LazyLock, Mutex};
 
     /// Serializes tests that manipulate `NORA_CURATION_*` env vars.
@@ -3104,6 +1008,7 @@ mod tests {
 
     #[test]
     fn test_env_override_storage() {
+        let _lock = ENV_MUTEX.lock().unwrap();
         let mut config = Config::default();
         std::env::set_var("NORA_STORAGE_MODE", "s3");
         std::env::set_var("NORA_STORAGE_PATH", "/data/nora");
@@ -3122,6 +1027,7 @@ mod tests {
 
     #[test]
     fn test_env_override_storage_mode_local() {
+        let _lock = ENV_MUTEX.lock().unwrap();
         let mut config = Config::default();
         std::env::set_var("NORA_STORAGE_MODE", "local");
         config.apply_env_overrides().unwrap();
@@ -3131,6 +1037,7 @@ mod tests {
 
     #[test]
     fn test_env_override_storage_mode_case_insensitive() {
+        let _lock = ENV_MUTEX.lock().unwrap();
         let mut config = Config::default();
         std::env::set_var("NORA_STORAGE_MODE", "S3");
         config.apply_env_overrides().unwrap();
@@ -3140,6 +1047,7 @@ mod tests {
 
     #[test]
     fn test_env_override_storage_mode_rejects_unknown() {
+        let _lock = ENV_MUTEX.lock().unwrap();
         let mut config = Config::default();
         std::env::set_var("NORA_STORAGE_MODE", "redis");
         let result = config.apply_env_overrides();
@@ -3653,6 +1561,7 @@ mod tests {
 
     #[test]
     fn test_config_file_sets_s3_mode_without_env() {
+        let _lock = ENV_MUTEX.lock().unwrap();
         // Regression test for issue #4: config.toml mode="s3" must work
         // without NORA_STORAGE_MODE env var (previously overridden by
         // Dockerfile ENV NORA_STORAGE_MODE=local)
